requirements.txt
APScheduler==3.6.3
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
decorator==4.4.2
environs==9.2.0
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.2
marshmallow==3.9.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.20
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
